Move build_full_slice_tuple to unsage intrinsics and test
